Fixed version sorting; moved version functions to separate file
