+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
+# Unreleased
+
+* Fixed task version sorting issue (\#733);
+
 # 1.16.1
 
 * Improved handling of new default values in task version update (\#731);
