+++ components/src/lib/index.js
@@ -1,7 +1,12 @@
 import JSchema from './jschema/JSchema.svelte';
 import PropertyDescription from './jschema/properties/PropertyDescription.svelte';
 import { SchemaValidator, detectSchemaVersion } from './jschema/jschema_validation';
-import { deepCopy, getValidationErrorMessage, stripNullAndEmptyObjectsAndArrays } from './utils';
+import {
+	deepCopy,
+	getValidationErrorMessage,
+	stripNullAndEmptyObjectsAndArrays
+} from './common/utils';
+import { greatestVersionDesc, greatestVersionAsc } from './common/version';
 import { getPropertiesToIgnore } from './jschema/property_utils';
 import { JsonSchemaDataError } from './jschema/form_manager';
 import { stripIgnoredProperties } from './jschema/jschema_adapter';
@@ -19,5 +24,7 @@ export {
 	getValidationErrorMessage,
 	PropertyDescription,
 	detectSchemaVersion,
-	FilteredTasksTable
+	FilteredTasksTable,
+	greatestVersionDesc,
+	greatestVersionAsc
 };
